Move safe_join to utils
